refactor: Логика получения данных для страницы offer вынесена в хук useOfferInfo
